modify type int -> long

+++ shared/src/commonMain/kotlin/com/kmm/moonflower/feature/garden/data/local/GardenPlantingLocalDataSource.kt
@@ -8,7 +8,7 @@ interface GardenPlantingLocalDataSource {
 
     suspend fun isExistPlantInGardenPlanting(plantId: String): Boolean
 
-    suspend fun insertGardenPlanting(plantId: String, plantDate: Int, lastWateringDate: Int)
+    suspend fun insertGardenPlanting(plantId: String, plantDate: Long, lastWateringDate: Long)
 
     suspend fun deleteGardenPlanting(gardenId : Int)
 }
