+++ shared/src/commonMain/kotlin/com/kmm/moonflower/feature/garden/data/repository/GardenPlantingRepositoryImpl.kt
@@ -17,8 +17,8 @@ class GardenPlantingRepositoryImpl(
 
     override suspend fun insertGardenPlanting(
         plantId: String,
-        plantDate: Int,
-        lastWateringDate: Int
+        plantDate: Long,
+        lastWateringDate: Long
     ) {
         local.insertGardenPlanting(
             plantId = plantId,
